# swagger.yaml
openapi: 3.0.0
info:
    title: Medical Booking API
    description: Made by - P's Get Degrees
    contact:
        name: Go to the Front End
        url: https://api-medical-booking.herokuapp.com/
    version: "1.0"
servers:
    - url: https://api-medical-booking.herokuapp.com/api/v1
      variables: {}
paths:
...
    /user:
        get:
            tags:
                - user
            summary: Get All Users
            operationId: GetAllUsers
            parameters:
                - name: role
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [patient, doctor, admin]
                      example: patient
                - name: name
                  in: query
                  required: false
                  schema:
                      type: string
                      pattern: "^[A-Za-z-]+$"
                - name: surname
                  in: query
                  required: false
                  schema:
                      type: string
                      pattern: "^[A-Za-z-]+$"
                - name: email
                  in: query
                  required: false
                  schema:
                      type: string
                - name: gender
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [other, male, female]
                      example: other
                - name: dateFilters
                  in: query
                  required: false
                  description: "Possible variables ( createdAt , updatedAt )  |  Supported operators ( > , >= , = , < , <= ) "
                  schema:
                      type: string
                      example: "createdAt>=1633159914979"
                - name: sort
                  in: query
                  required: false
                  description: "This will sort the returning items in either numeric or alphabetic and can be inversed by adding a ( - ) operator prefixing the field. Sort filters can be chained using the ( , ) operator"
                  schema:
                      type: string
                      example: "name,-surname"
                - name: select
                  in: query
                  required: false
                  description: "By default the query return all fields on the schema, defining a select field(s) will return on these field(s). Select filters can be chained using the ( , ) operator"
                  schema:
                      type: string
                      example: "name,surname,email,mobile"
                - name: page
                  in: query
                  required: false
                  description: "By default the return data is segmetented depending on the limit of return and will return the first segment on the database. If the page value is 5 it will return the 5th segment of the database"
                  schema:
                      type: string
                      pattern: "^[0-9]+$"
                - name: limit
                  in: query
                  required: false
                  description: "By default the return data is 10 items. Setting the limit field to 100 will return 100 items."
                  schema:
                      type: string
                      pattern: "^[0-9]+$"
            responses:
                "200":
                    description: "Logged in Successfully"
                    headers: {}
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                _id:
                                                    type: string
                                                    example: "61766a7db8d4b12febe50d28"
                                                email:
                                                    type: string
                                                    example: "aazaan@gmail.com"
                                                password:
                                                    type: string
                                                    example: "$2a$10$rjblTeJZV9FLkR.g2GXheeIZkXwlcuM0jFhCio03sMhYWfJmAIPWG"
                                                role:
                                                    type: string
                                                    example: doctor
                                                name:
                                                    type: string
                                                    example: aazaan
                                                surname:
                                                    type: string
                                                    example: kemp
                                                gender:
                                                    type: string
                                                    example: other
                                                nationality:
                                                    type: string
                                                    example: Croatian
                                                mobile:
                                                    type: string
                                                    example: "0331432087"
                                                address:
                                                    type: string
                                                    example: "919 imagination court, south australia"
                                                createdAt:
                                                    type: string
                                                    example: "2021-10-25T08:27:41.979Z"
                                                updatedAt:
                                                    type: string
                                                    example: "2021-10-25T08:27:41.979Z"
                                                __v:
                                                    type: number
                                                    example: 0
                                    count:
                                        type: string
                                        example: "1"
            deprecated: false
    /user/{id}:
        parameters:
            - in: path
              name: id
              schema:
                  type: string
              required: true
              description: The user ID
        get:
            tags:
                - user
            summary: Get User
            operationId: GetUser
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
        patch:
            tags:
                - user
            summary: Update Users
            operationId: UpdateUser
            parameters: []
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        encoding: {}
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    example: user@gmail.com
                                    pattern: '^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$'
                                password:
                                    type: string
                                    example: User1234!
                                    description: "Must have minimum eight characters, at least one upper case English letter, one lower case English letter, one number and one special character"
                                    pattern: "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{8,}$"
                                name:
                                    type: string
                                    example: Robert
                                    minLength: 3
                                    maxLength: 50
                                    pattern: "^[A-Za-z-]+$"
                                surname:
                                    type: string
                                    example: Alexander
                                    minLength: 3
                                    maxLength: 50
                                    pattern: "^[A-Za-z-]+$"
                                gender:
                                    type: string
                                    example: other
                                    enum: [other, male, female]
                                mobile:
                                    type: string
                                    description: "Must start with 02, 04, 03, 07, 08 and have 10 digits all together"
                                    example: "0412345678"
                                    pattern: '^\({0,1}((0|\+61)(2|4|3|7|8)){0,1}\){0,1}(\ |-){0,1}[0-9]{2}(\ |-){0,1}[0-9]{2}(\ |-){0,1}[0-9]{1}(\ |-){0,1}[0-9]{3}$'
                                address:
                                    type: string
                                    example: "123 Testing Street, NSW 2000"
                                    minLength: 3
                                    maxLength: 100
                required: false
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
...
components:
    securitySchemes:
        httpBearer:
            type: http
            scheme: bearer
security:
    - httpBearer: []
...